Create user/pass authentication and login screen
The user/pass authentication is part of the Auth module. It takes the
username and passwords provided, checks that the user exists, checks
that the password is correct(using comeonin), and either returns a
logged in user, or an error.

The sessions controller handles the creation of these sessions, using
the Auth module. There is a corresponding view and new template for this
controller where the user can login.