Create and delete videos for user
Most of the business logic is in the categories context. So calling
Categories.delete_video_for_user with the user passed in will look for
the association, then only delete the video if it's associated.

Currently, videos are still associated with users, but as this is
cross-context, I am not sure whether this is the correct way to do this
in Phoenix 1.3.

The video deletion only matches on the {:ok, _video} result at the
moment. Need to add error handling or fallback controller to handle
failed video deletion. Or just show the correct videos to the correct
users.