Create belongs_to relationship for videos topics
The phoenix generator does not automatically generate belongs_to
relationships, instead adding IDs to the schema. The user_id and
topic_id have been removed to be replaced with belongs_to clauses.

The videos table has the topic reference added. This involves adding a
reference on the topic_id, and creating an index for this foreign key.