List topics by name and id, in alphabetical order
When Categories.list_topics_by_name_and_id is called, it builds a query,
which filters by name and ID, and makes order alphabetical. This query
is built using functions in the Categories.Query module.

The query is then called, and topics are returned.

The Context/Schema/Query structure is based on this:
https://robots.thoughtbot.com/lessons-from-using-phoenix-1-3